Move theme slice to settings, useSettings hook update
