+++ src/store/settings.ts
@@ -1,12 +1,15 @@
 import { createSelector, createSlice } from '@reduxjs/toolkit';
+import { AppTheme } from 'types';
 import { RootState } from '.';
 
 type State = {
   soundEnabled: boolean;
+  theme: AppTheme;
 };
 
 const initialState: State = {
   soundEnabled: true,
+  theme: 'light',
 };
 
 const settingsSlice = createSlice({
@@ -19,16 +22,25 @@ const settingsSlice = createSlice({
     disableSound: (state) => {
       state.soundEnabled = false;
     },
+    enableDarkMode: (state) => {
+      state.theme = 'dark';
+    },
+    disableDarkMode: (state) => {
+      state.theme = 'light';
+    },
   },
 });
 
 export const settings = settingsSlice.reducer;
 
-export const { enableSound, disableSound } = settingsSlice.actions;
+export const { enableSound, disableSound, enableDarkMode, disableDarkMode } =
+  settingsSlice.actions;
 
-const selectState = (state: RootState) => state.preferences;
+const selectState = (state: RootState) => state.settings;
 
 export const selectIsSoundEnabled = createSelector(
   [selectState],
   (state) => state.soundEnabled,
 );
+
+export const selectTheme = createSelector(selectState, (state) => state.theme);
